Renamed package - Required refactor of all appropriate packages - Required clean and rebuild - Required gradle sync
