+++ app/src/androidTest/java/com/example/pixelpedia/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.example.pixelpedia
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
